requirements.txt
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.2.0
aiohttp==3.9.1
PyPDF2==3.0.1
pdfplumber==0.10.3
